Sp playground settings (#970)
This changes the playground to work with new flat options.
It converts all the tests in the gallery, lib, and playground (e2e and units) to new flat options.
It makes the blueprint formatting all support the flat options through a new { newSPs: true} option. (forced in the playground to support this change.